Minor changes. Fixed error introduced in last update
Error fixed in the last updated that resulted in >98% success. Played around with the features and other values a bit. Made it so it is much easier to update the features. Teacher asked for me to find one of the seeds that results in a 90+% success rate, so I did.